sonata review suggestion +bottomsheet

+++ features/top_ads/top_ads_dashboard/src/main/java/com/tokopedia/topads/dashboard/view/adapter/insight/TopadsRecomGroupBsAdapter.kt
@@ -71,7 +71,7 @@ class TopadsRecomGroupBsAdapter(private val onGroupSelect: (() -> Unit)) : Recyc
     }
 
     fun getCheckedPosition(): Int {
-        return items[selectedPosition].groupId
+        return items[selectedPosition].groupId.toInt()
     }
 
     fun isChecked():Int{
