+++ features/top_ads/top_ads_management_common/src/main/java/com/tokopedia/topads/common/data/model/KeywordsItem.kt
@@ -12,7 +12,7 @@ open class KeywordsItem(
         @SerializedName("source")
         var source: String = "dashboard_add_product",
         @SerializedName("priceBid")
-        var priceBid: Int = 0,
+        var priceBid: Double = 0.0,
         @SerializedName("status")
         var status: String = "1"
 )
