+++ features/top_ads/top_ads_management_common/src/main/java/com/tokopedia/topads/common/data/response/GroupEditInput.kt
@@ -19,13 +19,13 @@ data class GroupEditInput(
             var adOperations: List<AdOperationsItem>? = null,
 
             @field:SerializedName("priceBid")
-            var priceBid: Int? = 0,
+            var priceBid: Double? = 0.0,
 
             @field:SerializedName("name")
             var name: String? = null,
 
             @field:SerializedName("priceDaily")
-            var dailyBudget: Int? = 0,
+            var dailyBudget: Double? = 0.0,
 
             @field:SerializedName("scheduleEnd")
             var scheduleEnd: String? = null,
